refactor(cli): move skill primitives out of the command into services
The skill-file primitives (paths, PLATFORM_TARGETS, read/write, target
prompt) lived in the skills install command, which extends BaseCommand.
Having BaseCommand's freshness check reach them created a circular
import (baseCommand -> skills -> install -> baseCommand) and a TDZ
crash during the AI-context build.

Move the framework-agnostic primitives and findStaleSkills into
services/skills.ts so the command, onboarding helpers, and BaseCommand
all depend downward on it. findStaleSkills also no longer lives under
onboarding, where it never belonged.
