+++ packages/cli/src/helpers/onboarding/skill-install.ts
@@ -2,7 +2,7 @@ import { dirname, relative } from 'path'
 import chalk from 'chalk'
 import prompts from 'prompts'
 
-import { PLATFORM_TARGETS, readSkillFile, writeSkillToTarget, formatPlatformName, promptForPlatformTarget } from '../../commands/skills/install.js'
+import { PLATFORM_TARGETS, readSkillFile, writeSkillToTarget, formatPlatformName, promptForPlatformTarget } from '../../services/skills.js'
 import { detectCliMode, detectOperator, OPERATOR_TO_PLATFORM } from '../cli-mode.js'
 import { makeOnCancel } from './prompts-helpers.js'
